+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Changed
 
 - Sync to upstream Luau 0.563
+- Prioritised common services and Instance properties/methods in autocomplete so that they show up first
 
 ### Fixed